main moved 10 commits (140 files) while this branch was in flight: a
long-history "Selection engine" ingest re-pulled full price history for all
18 AE names and all 11 SA names (~2.8x longer for most), a 15-year Samsung
ingest for KR (with an already-merged KR refit: nu Gaussian->12, cal
1.154->1.105), a GOLD refresh, and a roll-forward that graded the OCDI/ORHD/
GOLD/SAMSUNG T+20-T+60 cohorts and appended 3 new OCDI/ORHD sessions.

Conflicts (exactly the 3 GitHub flagged): engine/panel_hashes.json,
engine/panels/EG_OCDI_60d.csv, engine/panels/EG_ORHD_60d.csv. Resolved by
ground truth, not by picking a side: every conflicting hash was recomputed
directly from the current merged raw files, and both EG panels were rebuilt
via build_panel_file() against the merged OHLC. While doing this, main's own
panel_hashes.json conflict block turned out to carry stale hashes for
unrelated EG names (RMDA, CLHO, and likely more) -- the same staleness bug
this branch already fixed once for all 30 EG names pre-merge, reintroduced
independently on main. Recomputed rather than trusted.

Post-merge re-verification (engine/reverify_post_merge.py, log in
engine/reverify_log.txt): the original gate re-run numbers for AE, EG, KR,
SA, XAU were computed against libraries that no longer exist now that main's
ingest landed, so they were redone end to end for both tags (60d, 3m) on the
current merged raw data. IN/QA/US/XPT weren't touched by any concurrent
commit and keep their original numbers. This also rebuilt every AE/SA/KR/XAU
panel file and hash entry that had gone stale relative to the merged raw
CSVs (main's ingest updated raw data for AE/SA without rebuilding their
panels or registry) -- committed here for the same reason the two EG panels
were: an unrebuilt panel silently drifts from its own raw CSV.

Findings, all against today's live market_profiles.py incumbent:
- XAU, EG, KR: no new material trigger. EG is byte-identical (OCDI/ORHD's 3
  extra sessions moved nothing). KR's fresh 60d refit reproduces main's
  already-merged nu=12/cal=1.105 exactly, an incidental cross-check.
- AE 60d band moves -10.68% vs production on the grown library alone, with no
  horizon change involved -- this is main's own pending recalibration debt
  from the concurrent ingest, not something this branch introduces. Flagging
  it; not fixing it here, that's a separate promotion decision.
- SA is the one that matters for this PR: 60d band -8.08%, 3m band -7.68%,
  both over the 5% gate. SA/EXTRA's FAIL (the finding that motivated the
  adaptive-width work) reproduces at 3m on more than double the windows
  (392 vs 180) and is now also robust at 60d (was PARITY pre-ingest, FAIL
  post-ingest) -- more data sharpened the same conclusion, it did not
  reverse it. ALINMA moves PARITY->BOUNDARY at 60d. None of this is silently
  applied: market_profiles.py is untouched by this branch (verified, zero
  diff against origin/main), so production keeps running its current SA/AE
  fit until a human reviews and promotes a new one.

Net: the calendar-horizon infrastructure itself (engine/horizons.py,
mc_v3.calendar_horizons/carry_log_h/backtest_v3(horizon_months=), panel
namespacing, the site changes) is unaffected by any of this and remains
correct and bit-for-bit backward compatible. What changed is the supporting
evidence for two markets (AE, SA), surfaced honestly rather than left stale
in the PR.

+++ engine/data_quality.py
@@ -96,7 +96,13 @@ def clean_ohlc(df, ticker="", verbose=True, market=None):
     # (not interpolating) is deliberate: a session with no valid price contributes no
     # information, and synthesising one invents a trade that never happened.
     px = ['Price', 'Open', 'High', 'Low']
-    bad = (df[px] <= 0).any(axis=1) | ~np.isfinite(df[px].astype(float)).all(axis=1)
+    # Cast to float BEFORE any repair. The step-2 back-adjust multiplies a price
+    # column by a fractional factor; on an integer-dtype column (Korean prices are
+    # whole KRW, so pandas infers int64) that raises LossySetitemError and the gate
+    # DIES mid-repair. Never bit before because EG/AE/QA/SA/US/XAU prices all carry
+    # decimals and parse as float. Found 27-Jul-2026 on the Samsung 15-year ingest.
+    df[px] = df[px].astype(float)
+    bad = (df[px] <= 0).any(axis=1) | ~np.isfinite(df[px]).all(axis=1)
     if bad.any():
         dates = df['Date'][bad]
         log.append(f"dropped {int(bad.sum())} rows carrying a non-positive or "
